Made movement track layers of indirection from player push
Effectively, the player's own movement is zero layers of indirection. But, if they then push a log, that represents one layer of indirection. And if that log pushes over another log, that represents two layers of indirection. This is mostly useful when exploring raft movement - only movements relatively close to the player will cause rafts to sail.