Fixed a long-standing bug with entity deletion
Function RMEN is supposed to remove an entity from CENT, and return a pointer to the next entity in the list. Previously, the returned pointer was just whatever the next pointer of the deleted entity was. However, if the next entity in the linked list is moved into the deleted entity's spot in memory compaction, the returned pointer would then point to dead memory. This is now properly accounted for.