Got rid of PDAT, and tightened up sprite data in MESTAT
So, to walk through all the reasoning here, I firstly decided to get rid of postboxes. They're a bit of a technical burden, and given how small my maps are, they're just not very necessary. Given that, I no longer need the island discovery map in PDAT, which only leaves the snowfolk discovery map. But for how small of a feature that is (snowfolk brightening on discovery) it just didn't seem worthwhile to keep around. Altogether, that obsoletes all of PDAT!

But then, why also get rid of so many sprites? Well, I've also decided to get rid of flash rom save files. Again, my maps are rather small, so they're not hugely necessary. And more importantly, the creation and loading of those files would inevitably wear the flash rom out a little bit, which isn't preferable. So I've decided to get rid of them, which brings with it the snowfolk/island icons. Paired with the obsoleting of PDAT, we can chop off 7 sprites from the end of the entities sprite block.