Cleaned up entity type change logic
Previously, if the entity changed height, it would shift up or down with MVER to move other adjacent stuff, before then directly modifying its height in memory to correct its position. But this isn't very clean, especially for undo logic. As such, the entity now experiences no change in height itself, but instead looks for the lowest entity above it and moves that as necessary.