Simplified entity type change logic, and undoing of type change
Firstly, MVER is used in CHET, and I changed the undo to reflect the fact that the MVER will be logged separately. Additionally, I simplified the code around that aforementioned MVER. This also changes the behavior very slightly; previously if the type change changed the entity's height, other stuff would be placed directly on top of it. Now the height change is just added to the other stuff's height. This shouldn't matter, generally. But on the off chance the user finds a way to have an item hover over a log, they can keep it hovering a little longer!