Removed the ability to tumble logs up raft slopes
Funny enough, this was implemented in a very janky way in code: If a raft stopped sailing, then if it had something like a standing log on top, that standing log would fall over onto the opposite bank. This system was effectively hijacked to let you tumble logs up raft slopes.