Fixed bug in undoing raft bereavement
I originally coded it so that if a raft became a log, the removed log would always be added as the contained log, since that theoretically creates an equivalent raft. This works in the forward direction, but actually, subsequent undo steps may rely on the raft's main log having a particular ID. Thus the undo logic must properly recreate the raft as it was.