Corrected bug in undo pointer setting in CSAV
Said bug allows, effectively, arbitrary undo execution, and would very rarely come up in normal gameplay. I'm very tempted to leave it in. But I will remove it, on the off chance that some players like to hold down the undo button for a long while on a fresh save.